style changes, login/admin  check for links.

+++ Control/LinksControl.php
@@ -10,8 +10,17 @@ class LinksControl extends Control {
 	}
 
 	public function initialize() {
+		global $admin;
+		global $loggedIn;
+		$this->model->fail = FALSE;
 		if(isset($_POST['delete'])) {
-			$this->delete();
+			if($admin && $loggedIn) {
+				$this->delete();
+			} else {
+				$this->model->fail = TRUE;
+				$this->model->error = "<span class='error'>You're not logged in or you don`t have permissions to do this.</span>";
+			}
+
 		}
 		$this->model->links = EditableAreaDAO::selectByEditableAreaType(EditableArea::LINK);
 		$this->model->resourcePacks = EditableAreaDAO::selectByEditableAreaType(EditableArea::RESOURCE_PACK);
